Added lombok annotations in Product
