+++ src/main/java/com/es/phoneshop/model/entity/Product.java
@@ -1,0 +1,30 @@
+package com.es.phoneshop.model.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+import java.math.BigDecimal;
+import java.util.Currency;
+
+@Data
+@AllArgsConstructor
+public class Product {
+    private Long id;
+    private String code;
+    private String description;
+    /** null means there is no price because the product is outdated or new */
+    private BigDecimal price;
+    /** can be null if the price is null */
+    private Currency currency;
+    private int stock;
+    private String imageUrl;
+
+    public Product(String code, String description, BigDecimal price, Currency currency, int stock, String imageUrl) {
+        this.code = code;
+        this.description = description;
+        this.price = price;
+        this.currency = currency;
+        this.stock = stock;
+        this.imageUrl = imageUrl;
+    }
+}
